Try to define the S3 endpoint
This PR also starts to address usage of env macro (which is hardcoding
some strings in the binary that we want to keep dynamic).

This builds up on https://nickb.dev/blog/access-public-and-private-b2-s3-buckets-in-rust
But it seems to not work due to a visilibity change in the aws sdk: the config object is now private.